refactor: rename sharedui to ui
