+++ app/src/main/java/com/monstarlab/features/resources/ui/ResourceViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.monstarlab.core.extensions.onError
 import com.monstarlab.core.extensions.onSuccess
-import com.monstarlab.core.sharedui.errorhandling.ViewError
-import com.monstarlab.core.sharedui.errorhandling.mapToViewError
+import com.monstarlab.core.ui.errorhandling.ViewError
+import com.monstarlab.core.ui.errorhandling.mapToViewError
 import com.monstarlab.features.resources.domain.GetResourcesUseCase
 import com.monstarlab.features.resources.domain.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
